Fixed delete, and sub transform state access.
When I say fixed, I mean that the tests I wrote pass.
But I've written code that isn't being covered by these tests yet.

- I've made meta.state not actually use state but invoke the getter
of the prev path.  Thankfully that is cached after first invocation until
the next write, so shouldn't hurt performance too much.

- I've made meta.state return a list of results, not a single item.
This is how Z1-3 worked and I just somehow forgot to do it.
But it is necessary in order to have filter().name( 'Updated' ) work.

- I've changed the Op's meta to be their level, not the parents
but in order to do that I had to pass the parent meta into Property

- Filter remove dispatches to the prev path segments remove() so it
is sort of generic, but it special cases Traverse by editing the previous
state directly.  This is controversial, but I think, correct.