Not sure if this is 100% right yet
But here is the philosohpy:

For filter or values we augment the list itself not just the value
within it.  So when we first see a new reference in one
of those occasions, we capture a reference to the parent
and the index of our item within that parent.

If we subsequently drill down into someother part of the
state it shouldn't matter, we'll have a new reference
and we'll store the next index and parent.

When we get to the last op, there'll either be parents
or not.  If there is, it means there was a $values of $filter
somewhere before the final op.  We just loop through
the states, find the parent, and splice them out

If there ae no parent, it means this is the first traverse
so we can just set the list length to zero as we're
targeting all items in the original list.