Try to pass handler around more instead of proxy
I wanted to create an internal _get fn that the trap
would use, and the _internal would return the raw
handler not the proxy.  But it would lead to conditionals
which defeats the point.  Because the trap can return
different types of things so I can't simply grab the proxy
property returned from _get, I'd have to check what
type is returned and handle a few cases.

So keeping it as is basically for now, but prepping
for a future where I figure this out.